Refactor exercise file structure and modify README
This commit renames and moves an exercise file for consistency with the new 'k8s elements' file structure. It also adds a new exercise file. In addition, it simplifies the README, replacing detailed topic breakdowns in Curriculum Overview with the number of exercises. Corrections and improvements to the text have been made as well.