Add new Kubernetes exercise file and update README
The commit includes the addition of a new exercise file for kubectl commands and updates to the README. The new exercise file is meant to aid in the understanding and practice of Kubernetes primitives. This update also contains essential exercises related to setting and updating environment variables and container images in Kubernetes objects.