Update kubectl commands in exercise files
This commit updates several kubectl commands in the exercise files to correct syntaxes and simplify steps. Changes include saving an updated deployment file locally instead of applying it directly, correcting a label command, specifying that multiple pods should be updated with a label and a new environment variable, and modifying the way to retrieve and view logs from a pod. Minor changes were also made to how the labeling is performed in the "labels_exercise1" file.